Chore: Moved test files to their respective folders
