+++ packages/account/src/Components/poi/status/limited/__tests__/limited.spec.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { fireEvent, render, screen } from '@testing-library/react';
-import { POILimited } from '../poi-limited';
+import { POILimited } from '../limited';
 
 describe('<POILimited/>', () => {
     it('should render POILimited component', () => {
